Problem, solution, and task decomposition components
The directory "landing-page" has been renamed to "landingPage". All component file paths in HowItWorksData.tsx as well as the file paths of the components themselves have been updated accordingly to maintain consistency and coherence in the codebase.
