+++ components/landingPage/ImpactTrackerVisualizer.tsx
